WIP open discussion modal
