+++ apps/financiers/financiers/src/app/marketplace/title/modal/modal.component.html
@@ -1,0 +1,53 @@
+<global-modal *ngIf="movie$ | async as movie">
+  <ng-container *ngIf="campaign$ | async as camp">
+    {{ console(camp) }}
+  </ng-container>
+  <header>
+    <h2>What do you want to know about this project?</h2>
+  </header>
+
+  <section *ngIf="data.form as form">
+    <form [formGroup]="form" fxLayout="column"
+      (ngSubmit)="sendEmail(form.value, movie.title.international, data.orgs)">
+      <mat-form-field appearance="outline">
+        <mat-label>Subject</mat-label>
+        <mat-select formControlName="subject" required>
+          <mat-option value="main">Main</mat-option>
+          <mat-option value="artistic">Artistic</mat-option>
+          <mat-option value="additional">Additional</mat-option>
+          <mat-option value="financial">Financial</mat-option>
+          <mat-option value="investment">Investment</mat-option>
+        </mat-select>
+        <mat-error>Please select a subject</mat-error>
+      </mat-form-field>
+      <p class="mat-caption">
+        Please precise the amount you would like to invest.<br />
+        If you are not ready to provide us with such kind of information, just leave a message.
+      </p>
+      <div fxLayout fxLayoutGap="24px" formGroupName="scope">
+        <mat-form-field fxFlex appearance="outline">
+          <mat-label>From</mat-label>
+          <input type="number" matInput formControlName="from" min="0" />
+          <mat-icon matPrefix svgIcon="EUR"></mat-icon>
+        </mat-form-field>
+        <mat-form-field fxFlex appearance="outline">
+          <mat-label>To</mat-label>
+          <input type="number" matInput formControlName="to" min="0" />
+          <mat-icon matPrefix svgIcon="EUR"></mat-icon>
+        </mat-form-field>
+      </div>
+      <mat-form-field appearance="outline">
+        <mat-label>Message</mat-label>
+        <textarea required matInput formControlName="message"
+          placeholder="Questions about cast, financing, investment conditions..."></textarea>
+      </mat-form-field>
+      <div fxLayout="column" fxLayoutAlign="center center" fxLayoutGap="32px">
+        <button type="submit" mat-stroked-button color="primary" [disabled]="form.invalid">
+          <mat-icon svgIcon="thumb_up"></mat-icon>
+          <span>Send your message</span>
+        </button>
+        <button type="reset" mat-dialog-close mat-button>Cancel</button>
+      </div>
+    </form>
+  </section>
+</global-modal>
