+++ apps/financiers/financiers/src/app/marketplace/title/modal/modal.component.scss
@@ -1,0 +1,3 @@
+h2 {
+  text-align: center;
+}
