+++ apps/financiers/financiers/src/app/marketplace/title/view/view.component.ts
@@ -1,24 +1,16 @@
-﻿import { Component, ChangeDetectionStrategy, ViewChild, TemplateRef } from '@angular/core';
-import { getCurrencySymbol } from '@angular/common';
+﻿import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
-import { ActivatedRoute, Router } from '@angular/router';
-import { mainRoute, additionalRoute, artisticRoute, productionRoute } from '@blockframes/movie/marketplace';
-import { Organization } from '@blockframes/model';
-import { OrganizationService } from '@blockframes/organization/+state';
-import { CampaignService } from '@blockframes/campaign/+state';
-import { RouteDescription } from '@blockframes/utils/common-interfaces';
-import { SendgridService } from '@blockframes/utils/emails/sendgrid.service';
-import { templateIds } from '@blockframes/utils/emails/ids';
 import { MatDialog } from '@angular/material/dialog';
-import { pluck, shareReplay, switchMap, tap } from 'rxjs/operators';
-import { AuthService } from '@blockframes/auth/+state';
-import { UserService } from '@blockframes/user/+state';
-import { ErrorResultResponse } from '@blockframes/utils/utils';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { getUserEmailData, OrgEmailData } from '@blockframes/utils/emails/utils';
+import { ActivatedRoute, Router } from '@angular/router';
+import { CampaignService } from '@blockframes/campaign/+state';
+import { MovieCurrency, Organization } from '@blockframes/model';
 import { MovieService } from '@blockframes/movie/+state/movie.service';
-import { supportMailosaur } from '@blockframes/utils/constants';
-import { SnackbarErrorComponent } from '@blockframes/ui/snackbar/error/snackbar-error.component';
+import { additionalRoute, artisticRoute, mainRoute, productionRoute } from '@blockframes/movie/marketplace';
+import { OrganizationService } from '@blockframes/organization/+state';
+import { RouteDescription } from '@blockframes/utils/common-interfaces';
+import { combineLatest, forkJoin } from 'rxjs';
+import { pluck, shareReplay, switchMap, tap } from 'rxjs/operators';
+import { MarketplaceMovieModalComponent } from '../modal/modal.component';
 
 interface EmailData {
   subject: string;
@@ -34,8 +26,6 @@ interface EmailData {
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class MarketplaceMovieViewComponent {
-  @ViewChild('dialogTemplate') dialogTemplate: TemplateRef<unknown>;
-
   public movie$ = this.route.params.pipe(
     pluck('movieId'),
     switchMap((movieId: string) => this.movieService.getValue(movieId))
@@ -47,11 +37,10 @@ export class MarketplaceMovieViewComponent {
   );
 
   public campaign$ = this.movie$.pipe(
-    switchMap(movie => this.campaignService.valueChanges(movie.id)),
-    tap(campaign => this.currency = campaign.currency)
+    switchMap(movie => this.campaignService.valueChanges(movie.id))
   );
 
-  public currency: string;
+  public currency: MovieCurrency;
 
   public navLinks: RouteDescription[] = [
     mainRoute,
@@ -79,13 +68,9 @@ export class MarketplaceMovieViewComponent {
   constructor(
     private route: ActivatedRoute,
     private movieService: MovieService,
-    private authService: AuthService,
     private orgService: OrganizationService,
-    private userService: UserService,
     private campaignService: CampaignService,
     private dialog: MatDialog,
-    private snackBar: MatSnackBar,
-    private sendgrid: SendgridService,
     public router: Router
   ) { }
 
@@ -98,79 +83,13 @@ export class MarketplaceMovieViewComponent {
       }),
       message: new FormControl(),
     });
-    this.dialog.open(this.dialogTemplate, {
-      data: { orgs, form }
+    this.dialog.open(MarketplaceMovieModalComponent, {
+      data: {
+        orgs,
+        form,
+        movie: this.movie$,
+        campaign: this.campaign$
+      }
     });
-  }
-
-  async sendEmail(emailData: EmailData, title: string, orgs: Organization[]) {
-    try {
-      const templateId = templateIds.financiers.invest;
-      const userSubject = getUserEmailData(this.authService.profile);
-
-      const org = this.orgService.org;
-      const orgUserSubject: OrgEmailData = {
-        denomination: org.denomination.full ?? org.denomination.public,
-        id: org.id || '',
-        email: org.email || ''
-      }
-
-      const promises: Promise<ErrorResultResponse>[] = [];
-
-      const cyCheck = 'Cypress' in window;
-      let emailReady = false;
-      let numEmails = 0;
-
-      for (const org of orgs) {
-        const users = await this.userService.getValue(org.userIds);
-        for (const user of users) {
-          let toUser = getUserEmailData(user);
-          const userEmail = toUser.email;
-          // For e2e test purpose
-          if (cyCheck) {
-            toUser = { ...toUser, email: supportMailosaur };
-          }
-
-          const data = {
-            ...emailData,
-            currency: getCurrencySymbol(this.currency, 'wide'),
-            userSubject,
-            user: toUser,
-            org: orgUserSubject,
-            title,
-          };
-
-          /*
-           * If running E2E, for user other than sender,
-           * store it for access in E2E test.
-           * A single email is sufficient to check the email template
-           */
-          if (cyCheck && (userEmail !== userSubject.email)) {
-            emailReady = true;
-            ++numEmails;
-            window['cyEmailData'] = data;
-          }
-
-          const promise = this.sendgrid.sendWithTemplate({
-            request: { templateId, data, to: toUser.email },
-            app: 'financiers'
-          });
-          promises.push(promise);
-        }
-      }
-
-      if (cyCheck && emailReady) {
-        window['cyEmailData'].numEmails = numEmails;
-      }
-      const res = await Promise.all(promises);
-      const success = res.some(r => r.result);      
-      if (success) {
-        this.snackBar.open('Your email has been sent.', null, { duration: 3000 });
-      } else {
-        throw new Error('An error occured');
-       }
-    } catch (err) {
-      this.snackBar.openFromComponent(SnackbarErrorComponent, { duration: 5000 });
-    }
   }
 }
