+++ apps/financiers/financiers/src/app/marketplace/title/view/view.module.ts
@@ -2,12 +2,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
-import { ReactiveFormsModule } from '@angular/forms';
 import { FlexLayoutModule } from '@angular/flex-layout';
 
 // Component
 import { MarketplaceMovieViewComponent } from './view.component';
-import { SnackbarErrorModule } from '@blockframes/ui/snackbar/error/snackbar-error.module';
+import { MovieModalModule } from '../modal/modal.module';
 
 // Custom Modules
 import { MovieShellModule } from '@blockframes/movie/marketplace/shell/shell.module';
@@ -17,6 +16,7 @@ import { OrgChipModule } from '@blockframes/organization/components/chip/chip.mo
 import { PromotionalLinksModule } from '@blockframes/movie/components/promotional-links/promotional-links.module';
 import { WishlistButtonModule } from '@blockframes/organization/components/wishlist-button/wishlist-button.module';
 import { CampaignProgressModule } from '@blockframes/campaign/components/progress/progress.module';
+
 // Material
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
@@ -74,7 +74,6 @@ const routes = [
   declarations: [MarketplaceMovieViewComponent],
   imports: [
     CommonModule,
-    ReactiveFormsModule,
     FlexLayoutModule,
     MovieShellModule,
     ImageModule,
@@ -83,6 +82,7 @@ const routes = [
     WishlistButtonModule,
     PromotionalLinksModule,
     CampaignProgressModule,
+    MovieModalModule,
     // Material
     MatDialogModule,
     MatButtonModule,
@@ -91,7 +91,6 @@ const routes = [
     MatInputModule,
     MatSelectModule,
     MatSnackBarModule,
-    SnackbarErrorModule,
     // Routes
     RouterModule.forChild(routes)
   ]
